requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.3.0
click==8.1.4
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.6
flake8==6.0.0
inflection==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
uritemplate==4.1.1
